Create openchecker.js
this thing was originally built for UNM HSLIC and contains their hours. It can be modified without too much trouble to pull hours/holidays in from an external XML file or similar, but I had no server-side mechanism available to me to parse the data across domains and SOP kept the js from reading an XML file in another domain.